Fix bug which prevents passwordless sudo
The command to grant passwordless sudo permissions to the gravitysync user within the ssh container was failing to determine the correct line to modify to allow passwordless root. It was also set to read only.

By setting it to writable and using Regular Expressions to search for any line specifying the string "gravitysync", this has been fixed.